requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
rembg==2.0.59
pillow==11.1.0
python-multipart==0.0.20
onnxruntime==1.20.1
